wip... tearing things apart, making generic sidebar TODO: bottom items
