+++ airbyte-webapp/src/views/layout/MainView/SideBar.module.scss
@@ -1,0 +1,1 @@
+@forward "../SideBar/SideBar.module.scss";
