+++ airbyte-webapp/src/views/layout/MainView/SideBar.tsx
@@ -1,0 +1,26 @@
+import { FlexContainer } from "components/ui/Flex";
+
+import styles from "./SideBar.module.scss";
+import { GenericSideBar } from "../SideBar/GenericSideBar";
+import { MainNav } from "../SideBar/MainNav";
+import { TopItems } from "../SideBar/TopItems";
+
+export const SideBar: React.FC = () => {
+  return (
+    <GenericSideBar>
+      <FlexContainer direction="column">
+        <TopItems />
+        <FlexContainer
+          direction="column"
+          alignItems="center"
+          justifyContent="space-between"
+          className={styles.menuContent}
+        >
+          <TopItems />
+          <MainNav />
+          {/* <BottomItems /> */}
+        </FlexContainer>
+      </FlexContainer>
+    </GenericSideBar>
+  );
+};
