+++ airbyte-webapp/src/views/layout/SideBar/BottomItems.module.scss
@@ -1,0 +1,1 @@
+@forward "./SideBar.module.scss";
