+++ airbyte-webapp/src/views/layout/SideBar/BottomItems.tsx
@@ -1,0 +1,29 @@
+export const BottomItems: React.FC = ({ children }) => {
+  return <div />;
+  //     <FlexContainer direction="column" gap="xs" className={styles.bottomItems}>
+  //       const OSSBottomMenuItems = [
+  //   <a href={links.updateLink} target="_blank" rel="noreferrer" className={styles.menuItem} data-testid="updateLink">
+  //     <FontAwesomeIcon className={styles.helpIcon} icon={faRocket} />
+  //     <Text className={styles.text} size="sm">
+  //       <FormattedMessage id="sidebar.update" />
+  //     </Text>
+  //   </a>,
+
+  //   </DropdownMenu>,
+  //   <NavLink className={navLinkClassName} to={RoutePaths.Settings}>
+  //     <React.Suspense fallback={null}>
+  //       <NotificationIndicator />
+  //     </React.Suspense>
+  //     <SettingsIcon />
+  //     <Text className={styles.text} size="sm">
+  //       <FormattedMessage id="sidebar.settings" />
+  //     </Text>
+  //   </NavLink>,
+  // ];
+
+  // if (config.version) {
+  //   OSSBottomMenuItems.push(<Version primary />);
+  // }
+  //     </FlexContainer>
+  //   );
+};
