+++ airbyte-webapp/src/views/layout/SideBar/MainNav.tsx
@@ -1,53 +1,42 @@
 import { FormattedMessage } from "react-intl";
-import { NavLink } from "react-router-dom";
 
 import { FlexContainer } from "components/ui/Flex";
-import { Text } from "components/ui/Text";
 
-import { useExperiment } from "hooks/services/Experiment";
 import { RoutePaths } from "pages/routePaths";
 
 import ConnectionsIcon from "./components/ConnectionsIcon";
+import { CustomNavLink } from "./components/CustomNavLink";
 import DestinationIcon from "./components/DestinationIcon";
 import SourceIcon from "./components/SourceIcon";
-import { useCalculateSidebarStyles } from "./SideBar";
+import styles from "./MainNav.module.scss";
 
 export const MainNav: React.FC = () => {
-  const navLinkClassName = useCalculateSidebarStyles();
-  const showBuilderNavigationLinks = useExperiment("connectorBuilder.showNavigationLinks", false);
-
   return (
     <ul data-testid="navMainItems">
-      <FlexContainer direction="column" gap="sm">
+      <FlexContainer direction="column" gap="sm" className={styles.menuContent}>
         <li>
-          <NavLink className={navLinkClassName} to={RoutePaths.Connections} data-testid="connectionsLink">
-            <FlexContainer direction="column" alignItems="center" justifyContent="space-between">
-              <ConnectionsIcon />
-              <Text size="sm">
-                <FormattedMessage id="sidebar.connections" />
-              </Text>
-            </FlexContainer>
-          </NavLink>
+          <CustomNavLink
+            label={<FormattedMessage id="sidebar.connections" />}
+            icon={<ConnectionsIcon />}
+            to={RoutePaths.Connections}
+            testId="connectionsLink"
+          />
         </li>
         <li>
-          <NavLink className={navLinkClassName} to={RoutePaths.Source} data-testid="sourcesLink">
-            <FlexContainer direction="column" alignItems="center" justifyContent="space-between">
-              <SourceIcon />
-              <Text size="sm">
-                <FormattedMessage id="sidebar.sources" />
-              </Text>
-            </FlexContainer>
-          </NavLink>
+          <CustomNavLink
+            label={<FormattedMessage id="sidebar.sources" />}
+            icon={<SourceIcon />}
+            to={RoutePaths.Source}
+            testId="sourcesLink"
+          />
         </li>
         <li>
-          <NavLink className={navLinkClassName} to={RoutePaths.Destination} data-testid="destinationsLink">
-            <FlexContainer direction="column" alignItems="center" justifyContent="space-between">
-              <DestinationIcon />
-              <Text size="sm">
-                <FormattedMessage id="sidebar.destinations" />
-              </Text>
-            </FlexContainer>
-          </NavLink>
+          <CustomNavLink
+            label={<FormattedMessage id="sidebar.destinations" />}
+            icon={<DestinationIcon />}
+            testId="destinationsLink"
+            to={RoutePaths.Destination}
+          />
         </li>
       </FlexContainer>
     </ul>
