+++ airbyte-webapp/src/views/layout/SideBar/SideBar.module.scss
@@ -29,48 +29,39 @@
   width: 100%;
 }
 
-.bottomMenu {
-  padding: 0;
-  width: 100%;
-}
+// .helpIcon {
+//   font-size: 21px;
+//   line-height: 21px;
+// }
 
-.text {
-  margin-top: 7px;
-}
+// .dropdownMenuButton {
+//   display: flex;
+//   flex-direction: column;
+//   align-items: center;
+//   justify-content: center;
+//   height: 70px;
+//   width: 100%;
+//   background: transparent;
+//   border: 0;
+//   cursor: pointer;
+//   border-radius: variables.$border-radius-lg;
+//   color: inherit;
+//   text-decoration: none;
+//   transition: all variables.$transition-out;
 
-.helpIcon {
-  font-size: 21px;
-  line-height: 21px;
-}
+//   .text {
+//     margin-top: 7px;
+//   }
 
-.dropdownMenuButton {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  height: 70px;
-  width: 100%;
-  background: transparent;
-  border: 0;
-  cursor: pointer;
-  border-radius: variables.$border-radius-lg;
-  color: inherit;
-  text-decoration: none;
-  transition: all variables.$transition-out;
+//   &.open {
+//     background-color: colors.$blue-100;
+//   }
 
-  .text {
-    margin-top: 7px;
-  }
-
-  &.open {
-    background-color: colors.$blue-100;
-  }
-
-  &:hover:not(.open),
-  &:focus-visible:not(.open) {
-    background-color: colors.$grey-50;
-  }
-}
+//   &:hover:not(.open),
+//   &:focus-visible:not(.open) {
+//     background-color: colors.$grey-50;
+//   }
+// }
 
 .menuItem {
   background-color: transparent;
